info: Give more control to users about how to print objects
Currently using GST_DEBUG_OBJECT only the name of the object is printed,
but in environments where many pipelines are running concurrently it
quickly becomes hard to follow what a name refers to as several
elements/pads can have the same name in those pipeline.

Implementation:
---------------

This adds a GstObjectRepr private mini object (so it is ref counted) to
GstObject which contains needed information about how to quickly build
path to an object in its hierarchy.

This structure is recreated recursively by GstBin whenever a
hierarchical change is made within it (child added/removed/ etc...).

To avoid needing locking, we create a new GstObjectRepr object as previous
one can be used somewhere else at the same time (meaning it owns another
ref on it).

Usage:
------

This patch adds 2 options to the 'GST_DEBUG_OPTIONS' env variable
"objects-path" and "objects-toplevel" which makes the debug system
print the full object path in the hierarchy or only its toplevel
bin.

---
Note that the tests are only built with meson has they require
access to internal function which is only possible thanks
to the meson extract_objects feature.

https://bugzilla.gnome.org/show_bug.cgi?id=773463