WIP- ADD INTEGRATION TESTS - pipeline-pool-manager: Make video decoders output frames that are out of segment
And let `videorate` properly drop the ones that are not needed.
This is necessary for streams where there are gaps in the stream where
we should consider missing frames as "duplicate previous ones", for
examples, when you have frames like that in a stream:

```
[0   5]          [10   50]         [150 180]
```

if the users wants to output [60  100] for example, no frame will be
used while the logical behavior is to duplicate the frame [10-50] to
fill the segment instead of not showing any frame. With that patch
it is what happens.