TELLA: multiarch build on github actions (#2)
* ges: asset: Avoid trying to load twice the same asset

When requesting an asset from different threads we had no
guarantee that during the time we lookup an asset (which didn't exist)
and the time we create the asset with the same type/ID another thread
could not end up doing the same thing. In turns we could end up with
2 different threads loading the exact same asset and the cache
basically forgetting about one of the entries meaning that the user
would never get notified about one of those being ready to be used.

There was also the case when requesting "sync" where the user was
requesting an asset while another thread is creating it so it was
still in "ASSET_INITIALIZING" state, meaning that the returned asset
would be NULL which would be considered as an error in apps.

Since the cache lock is recursive we can just take it during the whole
ges_asset_request_async call and have other method still hold it.

* framepositioner: Make zorder controllable and expose it

* trying multiarch build on github actions

* aws cli for amd64

* dont fail on nonexistant tag

* removed unneeded needs

* added  --update because we got an error

* also run update here?

* run amd64 build on big runner

* trying to fix aac build with CFLAGS="-fPIC"

* no cacheg

* just a test

* try just the amd64 img

* update fdk-aac version?

* autogen

* added autoreconf

* its autoconf

* add automatke

* libtool

* cflags

* try fPIC

* -fPIC

* remove enable-pic

* disable shared

* enable-pic

* enable shared again, gstreamer needs it

* try this

* enable multiarch again

* runs-on

* push manifest multiarch

* tag sha

* dont forget to push

* arm as well

* do manifest push manually

* login first

* run gst-inspect-1.0 after building

* copy to root as well

* changed copy

* try again

* again

* try again 2

* trying

* install system fdkaac

* other libfdk-aac

---------

Co-authored-by: Thibault Saunier <tsaunier@igalia.com>